WIP - trying to make a 'fold' helper for transformers
So far the closest I've come to success is to use ScanLeft with some
wrapper state that also includes a 'toEmit' field. Then take everything
up through the first error result. At the moment there's an issue where
the fold doesn't emit its final result on EOF. That might be solved by
a mapConcatInput that converts EOF to a some signal value that tells the
fold to emit whatever's left.

It's really messy atm - just committing so I can work from a different
machine for a while.