more helpful error condition information
To elaborate:
Most handlers/parsers/transformers now have a `toString` override that
represents the combination of inner handlers/parsers/transformers that
compose them.
The GetFirstHandler will now throw an exception rather than returning a
result. This way the exception can be caught within the handler stack
and have some useful information attached to it.
The Splitter handler implementation will now attempt to catch exceptions
when passing events to its inner or downstream handlers. Such exceptions
will have a useful error message attached to them, which includes the
toString of the splitter's context matcher, and the handler which threw
the error.