Introduce Splitter.splitOnMatch, reimplement DefinitionList example
Ended up needing to parameterize the Splitter input type (which was
hard-coded to XMLEvent) in order to allow for splitters to work on
arbitrary streams.

Removed the FoldResults transformer because it should be equivalent to
ExpandResults >> Fold

Reworked most of the consumer handlers to no longer have Result output
types. They will now throw whatever errors they encounter. Added
`Consumer#safe` to wrap these consumers so that errors are caught as
Results. Parser (and Transformer.parse*) will use this under the hood.