refactor ContextMatcher to return Option instead of Result
Catching exceptions thrown by a ContextMatcher is now the responsibility
of the object that calls the match function, rather than the matcher
itself. It turns out that this only minorly affects the splitter
implementation, and a couple helper methods.

This change should make the `Result` class obsolete, with Parsers
using Try, and Matchers using Option. Removing the Result class will
come in the next commit.