WIP to make Parser more generic and not use Try anymore
I'm starting to make room in the interface for two classes to coexist:

 - XMLParser[A] for XMLEvents
 - JsonParser[A] for JSONEvent

This commit introduces another Stack management class in order to
generalize the Splitter handler, but I still need to reconcile it
with the existing Stackable typeclass.

Also renamed (just capitalization) all of the Transformer and Consumer
companion methods, and turned the corresponding case classes into
anonymous classes so they don't clutter the docs for Transformer
and Consumer.

Parser[A] now no longer wraps results with Try. This results in things
being thrown directly out of the handler loop, but I think having the
Failure wrapper wasn't really gaining us anything useful. One test
had to be removed due to this - the one for catching "chooser" function
failures in the splitter's joiner function. It would make a substream
parse to a Failure if the chooser function threw, but I think it's
better to just put the onus on the client to protect their chooser.