Re-overhaul the internals (no more effects in core, handler refactor)
The end result is something between "pure functional" and the old design.
The pure Parser interface of `step` and `finish` is preserved, minus the
`F[_]` type, but moved into the Handler interface with the assumption
that a handler may be internally-mutable just like the old handlers.
Parser goes back to being a factory for Handlers (though a utility
Parser.Stateless is available for real "pure" Parsers).

Transformer gets a similar treatment, where the step/finish methods
are moved into the new `Transformer.Handler` trait, and Transformer acts
as a factory for those handlers.

The key aspect to this that allows for integrating with 3rd-party stream
libraries is that the `step` and `finish` functions return an emit/result
and a "next state", rather than having to call methods on some downstream
handler. This way, we can set up "driver" functions like stream or
iterator construction.