Make JsonErrorHandlingBehaviors to mimic the Xml one
Plugged the Jackson+JSON support library into that behavior suite
and found a few little oddities to fix, like how JsonStackFixer was
using the location of the incoming event as the location of the injected
"context pop" event, when it made more sense to use the location of the
previous event.

Also IntelliJ was freaking out about `JsonParser.apply[T]` when sugaring
away the 'apply', so I moved the `def apply` onto the `ParserApplyWithBoundInput`
class so it doesn't need to be added implicitly.