refactor transformer handlers ...again. For performance
This new model uses the concept of a "HandlerWrite" that gets passed
to the transformer handler's `push` and `finish` methods, with the
Signal abstraction as a way of saying "please stop pushing things to me".

For simplicity, I also removed the capability for a transformer to
"change states" by returning a different handler from its `step` method.
Now, a transformer's handler just has to be mutable.

As a result, chains of transformers should be able to directly call
each other, rather than emitting Chains of values and relying on an
interpreter to forward the output from one transformer to the input
of the next. Hopefully this reduced number of allocations will provide
performance more in line with the original spac handler design,
while also still allowing for the new features like fancy error handling
and conversion from Transformer to fs2.Pipe

Tests aren't updated yet, but the src/main stuff compiles.