Make ContextLocation a trait instead of a concrete class
The main point of this is so that the JavaxLocation wrapper logic can
create an instance of ContextLocation without actually reading the
location values and constructing an inner Map. Since every event pulled
from the javax XMLStreamReader had to do that, this ends up saving
a fairly significant amount of time (2 seconds off of what used to be
a 17-second test), which makes it somewhat official that the new SPaC
is *faster* than the old one (at least on that particular test).